+++ packages/engine/Source/Shaders/GlobeFS.glsl
@@ -83,6 +83,10 @@ in vec2 v_clippingPosition;
 flat in int v_regionIndex;
 #endif
 
+#ifdef SPLIT_TERRAIN
+uniform float u_terrainSplitDirection;
+#endif
+
 #if defined(GROUND_ATMOSPHERE) || defined(FOG) && defined(DYNAMIC_ATMOSPHERE_LIGHTING) && (defined(ENABLE_VERTEX_LIGHTING) || defined(ENABLE_DAYNIGHT_SHADING))
 uniform float u_minimumBrightness;
 #endif
@@ -335,6 +339,15 @@ void main()
         {
             discard;
         }
+#endif
+
+#ifdef SPLIT_TERRAIN
+    float splitPosition = czm_splitPosition;
+    if (u_terrainSplitDirection < 0.0 && gl_FragCoord.x > splitPosition) {
+        discard;
+    } else if (u_terrainSplitDirection > 0.0 && gl_FragCoord.x < splitPosition) {
+        discard;
+    }
 #endif
 
 #ifdef ENABLE_CLIPPING_PLANES